Restructure tests for utils, update imports
