+++ src/hooks/usePagination.ts
@@ -1,5 +1,5 @@
 import {useSearchParams} from 'react-router-dom';
-import {updateUrlParams} from '../utils/updateUrlParams/updateUrlParams';
+import {updateUrlParams} from '../utils/updateUrlParams';
 
 type UsePaginationReturnType = {
   page: number;
