+++ src/pages/MoviesListPage/MoviesListPage.tsx
@@ -2,8 +2,8 @@ import {useQuery, useQueryClient} from '@tanstack/react-query';
 import {useEffect} from 'react';
 import {useSearchParams} from 'react-router-dom';
 import {useDebounce} from '../../hooks/useDebounce';
-import {usePagination} from '../../hooks/usePagination.ts';
-import {useSearchInput} from '../../hooks/useSearchInput.ts';
+import {usePagination} from '../../hooks/usePagination';
+import {useSearchInput} from '../../hooks/useSearchInput';
 import {fetchMoviesList} from '../../api/movieApi/endpoints/fetchMoviesList';
 import {getErrorFallbackMessage} from '../../api/movieApi/utils/getErrorFallbackMessage';
 import ApiErrorScreen from '../../app/components/ApiErrorScreen/ApiErrorScreen';
