+++ src/pages/common/MovieCover/MovieCover.tsx
@@ -4,7 +4,7 @@ import {
   PlaceholderContainer,
   PlaceholderIcon,
 } from './styled';
-import {getImageUrl} from '../../../utils/getImageUrl/getImageUrl';
+import {getImageUrl} from '../../../utils/getImageUrl';
 
 type MovieCoverProps = {
   poster_path: string | null;
