+++ src/utils/getImageUrl.ts
@@ -1,5 +1,6 @@
-import {IMAGE_BASE_URL} from '../../api/movieApi/constants';
-import type {ImageSize} from './types';
+import {IMAGE_BASE_URL} from '../api/movieApi/constants';
+
+export type ImageSize = 'w500' | 'original';
 
 export const getImageUrl = (path: string | null, size: ImageSize = 'w500'): string | null => {
   if (!path) return null;
