--- src/utils/getImageUrl/types.ts
@@ -1,1 +1,0 @@
-export type ImageSize = 'w500' | 'original';
